+++ assets/db/data.sql
@@ -91,7 +91,8 @@ INSERT INTO `migration` (`version`, `apply_time`) VALUES
   ('m220528_175811_remove_user_privilege_tables', '1653760931'),
   ('m220710_080845_remove_odt_templates', '1657440574'),
   ('m220710_114056_document_file_groups', '1657453642'),
-  ('m220730_144556_voting_block_settings_usergroup_order', '1659192616')
+  ('m220730_144556_voting_block_settings_usergroup_order', '1659192616'),
+  ('m220806_131705_motion_modification_date', '1659792341')
 ;
 
 SET SQL_MODE = @OLD_SQL_MODE;